some changes * remove `expect` * move implementation from `build_help_aliases` to `collect_aliases`
